Initial index file reading logic
I could have done this using RS/RT and getline to read the file incrementally.
A trial of that method proved it worked, but a fair bit more difficult than
just stepping through the bytes of the file with substr. This tradeoff makes
the (correct) assumption that I don't care about index files so large they don't
fit in memory.

This commit is a working proof-of-concept but needs refactoring.